requirements.txt
streamlit
python-dotenv
supabase
openai
requests
email-validator